hybrid deserialiser to work with full dbz events not unwrapped ones
this is needed in order to be able to support the repartition actions correctly. otherwise
the information about the FK wouldn’t be available which is needed as partition key.

repartition topic creation checks which of the needed intermediate topics might have already
been created due to the launch of multiple instances